Refactor ISBN-registry publication request item view
