+++ src/frontend/components/isbn-registry/publicationRequests/IsbnPublicationRequestDataComponent.jsx
@@ -25,10 +25,13 @@
  *
  */
 
-import React, {useState, useEffect, useReducer} from 'react';
-import {FormattedMessage, useIntl} from 'react-intl';
-import {v4 as uuidv4} from 'uuid';
+import React, {useState, useEffect, useMemo, useReducer} from 'react';
 import PropTypes from 'prop-types';
+
+import {FormattedMessage, useIntl} from 'react-intl';
+import {useAuth} from 'react-oidc-context';
+
+import {v4 as uuidv4} from 'uuid';
 
 import {makeApiRequest} from '/src/frontend/actions';
 import useList from '/src/frontend/hooks/useList';
@@ -47,24 +50,29 @@ import DoneIcon from '@mui/icons-material/Done';
 import SaveIcon from '@mui/icons-material/Save';
 import ClearIcon from '@mui/icons-material/Clear';
 
+import useAppStateDispatch from '/src/frontend/hooks/useAppStateDispatch';
+
 import '/src/frontend/css/common.css';
 import '/src/frontend/css/requests/isbnIsmn/dataComponent.css';
 
 import ListComponent from '/src/frontend/components/common/ListComponent.jsx';
 import {PUBLICATION_TYPES} from '/src/frontend/components/common/form/constants';
-import SavePublisherModal from '/src/frontend/components/isbn-registry/subComponents/modals/SavePublisherModal/ModalComponent.jsx';
+import SavePublisherModal from '/src/frontend/components/isbn-registry/subComponents/modals/SavePublisherModal/index.jsx';
+
+import {publicationHasIdentifiers} from '/src/frontend/components/isbn-registry/publicationRequests/utils';
 
 function IsbnPublicationRequestDataComponent(props) {
   const {
-    currentRequest,
+    publicationRequest,
     setPublicationRequest,
-    isEdit,
-    authenticationToken,
-    hasIdentifiers,
-    setSnackbarMessage
+    isEdit
   } = props;
 
   const intl = useIntl();
+  const {user: {access_token: authenticationToken}} = useAuth();
+
+  const appStateDispatch = useAppStateDispatch();
+  const setSnackbarMessage = (snackbarMessage) => appStateDispatch({snackbarMessage});
 
   // Publisher selected from the list
   const [publisher, setPublisher] = useState({});
@@ -77,11 +85,12 @@ function IsbnPublicationRequestDataComponent(props) {
   // Loading publishers for autocomplete
   // Component state
   const initialSearchBody = {searchText: ''};
+  const hasIdentifiers = useMemo(() => publicationHasIdentifiers(publicationRequest), [publicationRequest]);
 
   const [searchBody, updateSearchBody] = useReducer((prev, next) => {
     // Refetch default when searchText has been emptied
     // Trigger autocomplete only after three or more characters
-    if(next.searchText.length > 3 || next.searchText.length === 0) {
+    if (next.searchText.length > 3 || next.searchText.length === 0) {
       return {...prev, ...next};
     }
 
@@ -116,7 +125,7 @@ function IsbnPublicationRequestDataComponent(props) {
   // Reset pristine information when request changes
   useEffect(() => {
     setAutocompleteIsPristine(true);
-  }, [currentRequest]);
+  }, [publicationRequest]);
 
   /* Non-editable fields depending on user's role */
   const isEditable = (key) => {
@@ -153,14 +162,14 @@ function IsbnPublicationRequestDataComponent(props) {
   /* Refreshes list from API */
   function updateSearchText(event, value, reason) {
     // During first event, set pristine to false
-    if(autocompleteIsPristine && event) {
+    if (autocompleteIsPristine && event) {
       setAutocompleteIsPristine(false);
     }
 
     // Manage autocomplete text state, search body state and publisher state
     // based on input type and value
     if (['input', 'reset'].includes(reason) && (value || value === '')) {
-      if(reason === 'reset') {
+      if (reason === 'reset') {
         setAutocompleteInput({label: value, value: null});
       } else {
         // Note: used for autocomplete state value getter evaluation
@@ -169,7 +178,7 @@ function IsbnPublicationRequestDataComponent(props) {
         setAutocompleteInput({label: value, value: true});
       }
 
-      if(value === '') {
+      if (value === '') {
         setPublisher({label: '', value: null});
       }
 
@@ -200,7 +209,7 @@ function IsbnPublicationRequestDataComponent(props) {
     const requestBody = {publisherId: null};
 
     const updatedRequest = await makeApiRequest({
-      url: `/api/isbn-registry/requests/publications/${currentRequest.id}/set-publisher`,
+      url: `/api/isbn-registry/requests/publications/${publicationRequest.id}/set-publisher`,
       method: 'PUT',
       values: requestBody,
       authenticationToken,
@@ -247,8 +256,8 @@ function IsbnPublicationRequestDataComponent(props) {
     if (isEdit) return genAuthors([...tempAuthorsArray, index], index + 1);
 
     if (
-      currentRequest[`lastName${index}`] !== '' &&
-      currentRequest[`firstName${index}`] !== null
+      publicationRequest[`lastName${index}`] !== '' &&
+      publicationRequest[`firstName${index}`] !== null
     ) {
       return genAuthors([...tempAuthorsArray, index], index + 1);
     }
@@ -257,12 +266,12 @@ function IsbnPublicationRequestDataComponent(props) {
   }
 
   // Getter for autocomplete value
-  function getAutocompleteValue(){
-    const requestPublisherValue = currentRequest.publisherId ? {label: currentRequest.publisherName, value: currentRequest.publisherId} : {label: '', value: null};
+  function getAutocompleteValue() {
+    const requestPublisherValue = publicationRequest.publisherId ? {label: publicationRequest.publisherName, value: publicationRequest.publisherId} : {label: '', value: null};
     const currentPublisherValue = publisher.value ? publisher : {label: '', value: null};
 
     // If user has not started using autocomplete, display current request value
-    if(autocompleteIsPristine) {
+    if (autocompleteIsPristine) {
       return requestPublisherValue;
     }
 
@@ -272,14 +281,14 @@ function IsbnPublicationRequestDataComponent(props) {
   }
 
   function autocompleteHasSavedValue() {
-    if(autocompleteIsPristine) {
+    if (autocompleteIsPristine) {
       return true;
     }
 
-    const requestPublisherValue = currentRequest.publisherId || null;
+    const requestPublisherValue = publicationRequest.publisherId || null;
     const currentPublisherValue = publisher.value || null;
 
-    if(currentPublisherValue === requestPublisherValue) {
+    if (currentPublisherValue === requestPublisherValue) {
       return true;
     }
 
@@ -364,7 +373,7 @@ function IsbnPublicationRequestDataComponent(props) {
                     isEdit ||
                     hasIdentifiers ||
                     !publisher.value ||
-                    publisher.value === currentRequest.publisherId
+                    publisher.value === publicationRequest.publisherId
                   }
                   className={hasIdentifiers ? '' : 'iconButton'}
                   title={intl.formatMessage({id: 'form.button.label.save'})}
@@ -393,7 +402,7 @@ function IsbnPublicationRequestDataComponent(props) {
             <div className="publisherInformationLink">
               <AccountBoxIcon />
               <Link
-                href={`/isbn-registry/publishers/${publisher.value ? publisher.value : currentRequest.publisherId}`}
+                href={`/isbn-registry/publishers/${publisher.value ? publisher.value : publicationRequest.publisherId}`}
                 target="_blank"
                 rel="noreferrer"
               >
@@ -403,52 +412,51 @@ function IsbnPublicationRequestDataComponent(props) {
           </div>
         </div>
         <SavePublisherModal
-          publicationRequest={currentRequest}
+          publicationRequest={publicationRequest}
           publisherId={publisher.value} // Note: autocomplete formatting maps publisher.id to publisher.value
           setPublicationRequest={setPublicationRequest}
           savePublisherModalOpen={savePublisherModalOpen}
           setSavePublisherModalOpen={setSavePublisherModalOpen}
           handleCloseSavePublisherModal={handleCloseSavePublisherModal}
-          {...props}
         />
         <ListComponent
           edit={isEdit && isEditable('publisherIdentifierStr')}
           fieldName="publisherIdentifierStr"
           label={<FormattedMessage id="form.common.publisherIdentifier" />}
-          value={currentRequest.publisherIdentifierStr ?? ''}
+          value={publicationRequest.publisherIdentifierStr ?? ''}
         />
         <ListComponent
           edit={isEdit && isEditable('officialName')}
           fieldName="officialName"
           label={<FormattedMessage id="form.common.name" />}
-          value={currentRequest.officialName}
+          value={publicationRequest.officialName}
         />
         <ListComponent
           edit={isEdit && isEditable('address')}
           fieldName="address"
           label={<FormattedMessage id="form.common.address" />}
-          value={currentRequest.address}
+          value={publicationRequest.address}
         />
         <ListComponent
           edit={isEdit && isEditable('zip')}
           fieldName="zip"
           label={<FormattedMessage id="form.common.zip" />}
-          value={currentRequest.zip}
+          value={publicationRequest.zip}
         />
         <ListComponent
           edit={isEdit && isEditable('city')}
           fieldName="city"
           label={<FormattedMessage id="form.common.city" />}
-          value={currentRequest.city}
+          value={publicationRequest.city}
         />
 
         { // Display locality only for dissertations
-          currentRequest.publicationType === PUBLICATION_TYPES.DISSERTATION && (
+          publicationRequest.publicationType === PUBLICATION_TYPES.DISSERTATION && (
             <ListComponent
               edit={isEdit && isEditable('locality')}
               fieldName="locality"
               label={<FormattedMessage id="request.publication.locality" />}
-              value={currentRequest.locality}
+              value={publicationRequest.locality}
             />
           )
         }
@@ -457,25 +465,25 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('phone')}
           fieldName="phone"
           label={<FormattedMessage id="form.common.phone" />}
-          value={currentRequest.phone}
+          value={publicationRequest.phone}
         />
         <ListComponent
           edit={isEdit && isEditable('contactPerson')}
           fieldName="contactPerson"
           label={<FormattedMessage id="form.common.contactPerson" />}
-          value={currentRequest.contactPerson}
+          value={publicationRequest.contactPerson}
         />
         <ListComponent
           edit={isEdit && isEditable('email')}
           fieldName="email"
           label={<FormattedMessage id="form.common.email" />}
-          value={currentRequest.email}
+          value={publicationRequest.email}
         />
         <ListComponent
           edit={isEdit && isEditable('langCode')}
           fieldName="langCode"
           label={<FormattedMessage id="request.publication.contactLanguage" />}
-          value={intl.formatMessage({id: `common.${currentRequest.langCode}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.langCode}`})}
         />
       </div>
 
@@ -488,53 +496,53 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('title')}
           fieldName="title"
           label={<FormattedMessage id="form.common.title" />}
-          value={currentRequest.title}
+          value={publicationRequest.title}
         />
         <ListComponent
           edit={isEdit && isEditable('subtitle')}
           fieldName="subtitle"
           label={<FormattedMessage id="form.common.subtitle" />}
-          value={currentRequest.subtitle}
+          value={publicationRequest.subtitle}
         />
         <ListComponent
           edit={isEdit && isEditable('language')}
           fieldName="language"
           label={<FormattedMessage id="request.publication.publicationLanguage" />}
           value={intl.formatMessage({
-            id: `common.${currentRequest.language?.toLowerCase()}`
+            id: `common.${publicationRequest.language?.toLowerCase()}`
           })}
         />
         <ListComponent
           edit={isEdit && isEditable('publicationMonth')}
           fieldName="publicationMonth"
           label={<FormattedMessage id="form.common.publicationMonth" />}
-          value={currentRequest.month}
+          value={publicationRequest.month}
         />
         <ListComponent
           edit={isEdit && isEditable('publicationYear')}
           fieldName="publicationYear"
           label={<FormattedMessage id="form.common.publicationYear" />}
-          value={currentRequest.year}
+          value={publicationRequest.year}
         />
-        {currentRequest.publicationIdentifierElectronical && (
+        {publicationRequest.publicationIdentifierElectronical && (
           <ListComponent
             edit={false} // Never editable directly
             fieldName="publicationIdentifierElectronical"
             format="electronical"
             label={<FormattedMessage id="request.publication.electronic" />}
-            value={currentRequest.publicationIdentifierElectronical}
+            value={publicationRequest.publicationIdentifierElectronical}
             cancelId={cancelIdentifier}
             removeId={removeIdentifier}
             token={authenticationToken}
           />
         )}
-        {currentRequest.publicationIdentifierPrint && (
+        {publicationRequest.publicationIdentifierPrint && (
           <ListComponent
             edit={false} // Never editable directly
             fieldName="publicationIdentifierPrint"
             format="print"
             label={<FormattedMessage id="request.publication.printed" />}
-            value={currentRequest.publicationIdentifierPrint}
+            value={publicationRequest.publicationIdentifierPrint}
             cancelId={cancelIdentifier}
             removeId={removeIdentifier}
             token={authenticationToken}
@@ -543,7 +551,7 @@ function IsbnPublicationRequestDataComponent(props) {
       </div>
 
       {/* Julkaisutoiminta - Publishing activities*/}
-      {currentRequest.publicationType !== PUBLICATION_TYPES.DISSERTATION && (
+      {publicationRequest.publicationType !== PUBLICATION_TYPES.DISSERTATION && (
         <div className="listComponentContainer">
           <Typography variant="h6" className="listComponentContainerHeader">
             <FormattedMessage id="form.common.publishingActivities" />
@@ -553,7 +561,7 @@ function IsbnPublicationRequestDataComponent(props) {
             fieldName="publishedBefore"
             label={<FormattedMessage id="request.publication.previouslyPublished" />}
             value={
-              currentRequest.publishedBefore
+              publicationRequest.publishedBefore
                 ? intl.formatMessage({id: 'common.yes'})
                 : intl.formatMessage({id: 'common.no'})
             }
@@ -563,9 +571,9 @@ function IsbnPublicationRequestDataComponent(props) {
             fieldName="publishingActivity"
             label={<FormattedMessage id="form.common.frequency" />}
             value={
-              currentRequest.publishingActivity
+              publicationRequest.publishingActivity
                 ? intl.formatMessage({
-                  id: `form.isbnIsmn.publishingActivities.option.${currentRequest.publishingActivity.toLowerCase()}`
+                  id: `form.isbnIsmn.publishingActivities.option.${publicationRequest.publishingActivity.toLowerCase()}`
                 })
                 : ''
             }
@@ -574,7 +582,7 @@ function IsbnPublicationRequestDataComponent(props) {
             edit={isEdit && isEditable('publishingActivityAmount')}
             fieldName="publishingActivityAmount"
             label={<FormattedMessage id="request.publication.publishingFrequency" />}
-            value={currentRequest.publishingActivityAmount}
+            value={publicationRequest.publishingActivityAmount}
           />
         </div>
       )}
@@ -588,55 +596,55 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('publicationFormat')}
           fieldName="publicationFormat"
           label={<FormattedMessage id="form.common.selectFormat" />}
-          value={intl.formatMessage({id: `form.isbnIsmn.format.option.${currentRequest.publicationFormat.toLowerCase()}`})}
+          value={intl.formatMessage({id: `form.isbnIsmn.format.option.${publicationRequest.publicationFormat.toLowerCase()}`})}
         />
         <ListComponent
           edit={isEdit && isEditable('printFormat')}
           fieldName="printFormat"
           label={<FormattedMessage id="form.common.printFormat" />}
-          value={currentRequest.type}
+          value={publicationRequest.type}
         />
         <ListComponent
           edit={isEdit && isEditable('typeOther')}
           fieldName="typeOther"
           label={<FormattedMessage id="publisherRegistry.publisher.typeOther" />}
-          value={currentRequest.typeOther}
+          value={publicationRequest.typeOther}
         />
         <ListComponent
           edit={isEdit && isEditable('printingHouse')}
           fieldName="printingHouse"
           label={<FormattedMessage id="publisherRegistry.publisher.manufacturer" />}
-          value={currentRequest.printingHouse}
+          value={publicationRequest.printingHouse}
         />
         <ListComponent
           edit={isEdit && isEditable('printingHouseCity')}
           fieldName="printingHouseCity"
           label={<FormattedMessage id="form.common.city" />}
-          value={currentRequest.printingHouseCity}
+          value={publicationRequest.printingHouseCity}
         />
         <ListComponent
           edit={isEdit && isEditable('copies')}
           fieldName="copies"
           label={<FormattedMessage id="publisherRegistry.publisher.run" />}
-          value={currentRequest.copies}
+          value={publicationRequest.copies}
         />
         <ListComponent
           edit={isEdit && isEditable('edition')}
           fieldName="edition"
           label={<FormattedMessage id="publisherRegistry.publisher.edition" />}
-          value={currentRequest.edition}
+          value={publicationRequest.edition}
         />
         <ListComponent
           edit={isEdit && isEditable('fileformat')}
           fieldName="fileformat"
           label={<FormattedMessage id="form.common.fileFormat" />}
-          value={currentRequest.fileformat}
+          value={publicationRequest.fileformat}
         />
         <ListComponent
           edit={isEdit && isEditable('fileformatOther')}
           fieldName="fileformatOther"
           label={<FormattedMessage id="publisherRegistry.publisher.fileFormatOther" />}
-          value={currentRequest.fileformatOther}
+          value={publicationRequest.fileformatOther}
         />
       </div>
 
@@ -648,7 +656,7 @@ function IsbnPublicationRequestDataComponent(props) {
         <ListComponent
           edit={isEdit && isEditable('comments')}
           fieldName="comments"
-          value={currentRequest.comments}
+          value={publicationRequest.comments}
         />
       </div>
 
@@ -667,19 +675,19 @@ function IsbnPublicationRequestDataComponent(props) {
                     edit={isEdit}
                     fieldName={`firstName${author}`}
                     label={<FormattedMessage id="request.publication.givenName" />}
-                    value={currentRequest[`firstName${author}`] ?? ''}
+                    value={publicationRequest[`firstName${author}`] ?? ''}
                   />
                   <ListComponent
                     edit={isEdit}
                     fieldName={`lastName${author}`}
                     label={<FormattedMessage id="request.publication.familyName" />}
-                    value={currentRequest[`lastName${author}`] ?? ''}
+                    value={publicationRequest[`lastName${author}`] ?? ''}
                   />
                   <ListComponent
                     edit={isEdit}
                     fieldName={`role${author}`}
                     label={<FormattedMessage id="publisherRegistry.publisher.role" />}
-                    value={currentRequest[`role${author}`]}
+                    value={publicationRequest[`role${author}`]}
                   />
                 </div>
               </div>
@@ -693,20 +701,20 @@ function IsbnPublicationRequestDataComponent(props) {
                   <Typography>
                     <FormattedMessage id="request.publication.givenName" />:
                   </Typography>
-                  <Typography>{currentRequest[`firstName${author}`]}</Typography>
+                  <Typography>{publicationRequest[`firstName${author}`]}</Typography>
                   <div>{author}</div>
                 </div>
                 <div className="authorsLastName">
                   <Typography>
                     <FormattedMessage id="request.publication.familyName" />:
                   </Typography>
-                  <Typography>{currentRequest[`lastName${author}`]}</Typography>
+                  <Typography>{publicationRequest[`lastName${author}`]}</Typography>
                 </div>
                 <div className="authorsRoles">
-                  {currentRequest[`role${author}`]?.map((item) => (
+                  {publicationRequest[`role${author}`]?.map((item) => (
                     <Chip
                       key={item}
-                      label={<FormattedMessage id={`form.isbnIsmn.authors.role.option.${item.toLowerCase()}`}/>}
+                      label={<FormattedMessage id={`form.isbnIsmn.authors.role.option.${item.toLowerCase()}`} />}
                     />
                   ))}
                 </div>
@@ -725,27 +733,27 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('publicationsPublic')}
           fieldName="publicationsPublic"
           label={<FormattedMessage id="request.publication.isPublic" />}
-          value={intl.formatMessage({id: `common.${currentRequest.publicationsPublic}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.publicationsPublic}`})}
         />
         <ListComponent
           edit={isEdit && isEditable('publicationsIntra')}
           fieldName="publicationsIntra"
           label={<FormattedMessage id="request.publication.publicationsIntra" />}
-          value={intl.formatMessage({id: `common.${currentRequest.publicationsIntra}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.publicationsIntra}`})}
         />
         <ListComponent
           edit={isEdit && isEditable('publicationType')}
           fieldName="publicationType"
           label={<FormattedMessage id="form.common.format" />}
-          value={intl.formatMessage({id: `common.${currentRequest.publicationType}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.publicationType}`})}
           publication="isbn-ismn"
         />
-        {currentRequest.publicationType === 'MAP' && (
+        {publicationRequest.publicationType === 'MAP' && (
           <ListComponent
             edit={isEdit && isEditable('mapScale')}
             fieldName="mapScale"
             label={<FormattedMessage id="form.common.scale" />}
-            value={currentRequest.mapScale}
+            value={publicationRequest.mapScale}
           />
         )}
       </div>
@@ -759,19 +767,19 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('volume')}
           fieldName="volume"
           label={<FormattedMessage id="form.common.volume" />}
-          value={currentRequest.volume}
+          value={publicationRequest.volume}
         />
         <ListComponent
           edit={isEdit && isEditable('series')}
           fieldName="series"
           label={<FormattedMessage id="form.common.title" />}
-          value={currentRequest.series}
+          value={publicationRequest.series}
         />
         <ListComponent
           edit={isEdit && isEditable('issn')}
           fieldName="issn"
           label={<FormattedMessage id="form.common.identifier" />}
-          value={currentRequest.issn}
+          value={publicationRequest.issn}
         />
       </div>
 
@@ -784,13 +792,13 @@ function IsbnPublicationRequestDataComponent(props) {
           edit={isEdit && isEditable('noIdentifierGranted')}
           fieldName="noIdentifierGranted"
           label={<FormattedMessage id="request.publication.noIdentifierGranted" />}
-          value={intl.formatMessage({id: `common.${currentRequest.noIdentifierGranted}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.noIdentifierGranted}`})}
         />
         <ListComponent
           edit={isEdit && isEditable('onProcess')}
           fieldName="onProcess"
           label={<FormattedMessage id="request.publication.onProcess" />}
-          value={intl.formatMessage({id: `common.${currentRequest.onProcess}`})}
+          value={intl.formatMessage({id: `common.${publicationRequest.onProcess}`})}
         />
       </div>
 
@@ -802,24 +810,24 @@ function IsbnPublicationRequestDataComponent(props) {
         <ListComponent
           edit={false}
           label={<FormattedMessage id="form.common.createdBy" />}
-          value={currentRequest.createdBy}
+          value={publicationRequest.createdBy}
         />
         <ListComponent
           edit={false}
           fieldName="timestamp"
           label={<FormattedMessage id="form.common.created" />}
-          value={currentRequest.created}
+          value={publicationRequest.created}
         />
         <ListComponent
           edit={false}
           label={<FormattedMessage id="form.common.modifiedBy" />}
-          value={currentRequest.modifiedBy}
+          value={publicationRequest.modifiedBy}
         />
         <ListComponent
           edit={false}
           fieldName="timestamp"
           label={<FormattedMessage id="form.common.modified" />}
-          value={currentRequest.modified}
+          value={publicationRequest.modified}
         />
       </div>
     </div>
@@ -827,12 +835,9 @@ function IsbnPublicationRequestDataComponent(props) {
 }
 
 IsbnPublicationRequestDataComponent.propTypes = {
-  currentRequest: PropTypes.object.isRequired,
+  publicationRequest: PropTypes.object.isRequired,
   setPublicationRequest: PropTypes.func.isRequired,
-  isEdit: PropTypes.bool,
-  authenticationToken: PropTypes.string.isRequired,
-  setSnackbarMessage: PropTypes.func.isRequired,
-  hasIdentifiers: PropTypes.bool.isRequired
+  isEdit: PropTypes.bool
 };
 
 export default IsbnPublicationRequestDataComponent;
